+++ rust/core-types/src/slice_error.rs
@@ -5,7 +5,7 @@
 use std::error::Error;
 use std::fmt::{Display, Formatter};
 
-/// Errors produced when normalizing a [`crate::subscript::Slice`].
+/// Errors produced when normalizing a [`crate::slice::Slice`].
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub enum SliceError {
 	StepZero,
